PhotoRename/requirements.txt
Pillow==9.5.0
piexif==1.1.3 